Indicate that sensors like 🌡️🔆⚖️ can be used
It's not very clear in the documentation that all these sensors are also supported and automatically discovered by home assistant with MQTT. Therefore, you find a lot of people who want this to be part of Esphome. I'll proceed to enrich the installation page that to enable the use of those sensors you need to select the ESP32-Sensors.bin file when flashing.