Added info regarding garmin devices.
In my testing i was able to make it work with this function, i only have this garmin device available so i pass this info to the community. If you have a garmin device try using nRF Connect (play store app) to find your device, if you don't find it use the Broadcast Heart Rate function.