Update apple.md
Updated to add a note regarding iOS devices stopping communications while the screen is off, while not universal this issue is very frustrating to solve if it is occurring. Users must work on the iOS device itself to try and get it to broadcast rather then focusing on ESPresense or home assistant.  Issue ESPresense #691